Fix: 이미지 URL이 올바르지 않으면 오류를 발생시키지 않고 이미지 엔티티만 삭제하도록 변경

+++ src/main/java/nbdream/image/service/ImageService.java
@@ -66,13 +66,13 @@ public class ImageService {
 
             String blobPath =request.imageUrl().replace(BASIC_PATH, "");
             Blob blob = storage.get(gcpStorageProperties.getBucketName(), blobPath);
-
-            Storage.BlobSourceOption preCondition =
-                    Storage.BlobSourceOption.generationMatch(blob.getGeneration());
-
-            boolean isDeleted = storage.delete(gcpStorageProperties.getBucketName(), blobPath, preCondition);
-            if(!isDeleted) {
-                throw new ImageDeleteFailException();
+            if (blob != null) {
+                Storage.BlobSourceOption preCondition =
+                        Storage.BlobSourceOption.generationMatch(blob.getGeneration());
+                boolean isDeleted = storage.delete(gcpStorageProperties.getBucketName(), blobPath, preCondition);
+                if(!isDeleted) {
+                    throw new ImageDeleteFailException();
+                }
             }
 
             Optional<Image> image = imageRepository.findByImageUrl(request.imageUrl());
